WIP: use 24 bit annotation

+++ framework/encode/capture_manager.h
@@ -148,8 +148,8 @@ class CaptureManager
     auto GetAllowPipelineCompileRequired() const { return allow_pipeline_compile_required_; }
 
     bool     IsAnnotated() const { return rv_annotation_info_.rv_annotation; }
-    uint16_t GetGPUVAMask() const { return rv_annotation_info_.gpuva_mask; }
-    uint16_t GetDescriptorMask() const { return rv_annotation_info_.descriptor_mask; }
+    uint64_t GetGPUVAMask() const { return rv_annotation_info_.gpuva_mask; }
+    uint64_t GetDescriptorMask() const { return rv_annotation_info_.descriptor_mask; }
     uint64_t GetShaderIDMask() const { return rv_annotation_info_.shaderid_mask; }
 
     uint64_t GetBlockIndex()
@@ -340,8 +340,8 @@ class CaptureManager
     struct
     {
         bool     rv_annotation{ false };
-        uint16_t gpuva_mask{ RvAnnotationUtil::kGPUVAMask };
-        uint16_t descriptor_mask{ RvAnnotationUtil::kDescriptorMask };
+        uint64_t gpuva_mask{ RvAnnotationUtil::kGPUVAMask };
+        uint64_t descriptor_mask{ RvAnnotationUtil::kDescriptorMask };
         uint64_t shaderid_mask{ RvAnnotationUtil::kShaderIDMask };
     } rv_annotation_info_;
 };
